Add Categories and Note components to JSONGenerator for better organization and modularity in the JSON editing interface
